requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.2
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
render==1.0.0
six==1.15.0
sqlparse==0.4.1